// src/app/shared/model/score-category/score-category-Schm-h-obj.model.ts
export class ScoreCategorySchmHObj {
    ScoreCategorySchmHId: number;
    ScoreCategorySchmHCode: string;
    ScoreCategorySchmHName: string;
    MrScoreTrxTypeCode: string;
    Notes: string;
    IsActive: boolean;
    ReservedField1: string;
    ReservedField2: string;
    ReservedField3: string;
    ReservedField4: string;
    ReservedField5: string;
    RowVersion: string;

    constructor() {
        this.RowVersion = "";
    }
}